Further context experimentation for v4
Cluster details are now stored identically to v3:

  ~/.cache/sre/clusters/v4/$CLUSTERID

Whereas kubeconfig files are stored separately in:

  ~/.config/sre/clusters/v4/$CLUSTERID

When running "source cluster $CLUSTERID", if no kubeconfig file is
present then "uhc cluster login" is run automatically to create one.

Would also like to run "uhc cluster login" if the login token in the
kubeconfig file is expired, but need to learn how to determine this.